Proof-of-concept linking with Firmament.
This change adds a new target ("test_integration") that links with
Firmament's libfirmament_scheduling.so (the scheduler library). In
addition, it also adds Firmament itself as an external dependency
and automatically builds the scheduler library.

Currently, running test_integration requires LD_LIBRARY_PATH to be
set explicitly unless libfirmament_scheduling.so is manually
copied to a system-wide library directory.

Change-Id: I6c2c3ad98c945b63e53e22d4efa9452987318187