Add Dockerfile and deployment script.
This adds a rudimentary script to create a Poseidon Docker image. The
script assumes that a full build has taken place prior to execution,
i.e., the Dockerfile does not itself build Poseidon from scratch.

Note that the script also does not currently perform any error checks
(e.g., whether the build targets exist) and that it only works on
Ubuntu 14.04 so far.

Change-Id: Ib853d10ec98992ac6099a4667c8aac7eb8ebd874