Check if container json resource object has requests object.
Poseidon is currently crashing if users submit pods that have
no resource requests. This is caused by an exception raised
when we parse json objects that contains the pods. This
change addresses makes sure that we don't blindly assume that
all pods have a resource or a request object.

Affected modules: apiclient

Change-Id: Ia7420997e26c5b08142f064fa83231e57743a577