Call scheduler events upon pod completion or failure.
Previously, Poseidon was not calling the Firmament scheduler
upon pod completion or failure. Thus, the scheduler would have
the impression that the pods are still running and would quickly
run out of resources.

This change handle such events. Moreover, it makes sure that
HandleTaskCompletion and HandleTaskFailure are only called the
first time Poseidon detects a pod has failed or completed.

Affected modules: firmament

Change-Id: I05e6169496a43275009ac547f2a36568a0bec0c8