Updated k8s_api_client to correctly parse resource units.
Previously, the k8s_api_client was assuming that memory units
are specified in KB and cpu units in multiples of CPU. However,
this is not the case in practice. This diff updates the code
to parse the different types of units that users can specify in
their config files.

Affected modules: apiclient

Change-Id: I618ccf5b3834ddb1db1c67eec5a3728d8218e0bd