Handle pods in unknown state.
This change handles the case when the pods are
in unknown state. This happens when a node fails. The
pods running on the failing node transition to the unknown
state. After a certain amount of time, Kubernetes replaces
these pods with new submitted pending pods. However, both
the pods in unknown and pending state can simulataneously be
in the system for a while.

Affected modules: firmament

Change-Id: If1ee1f655336ced4d0b2ad3017adeb59907db229