Adding NodeSelector/LabelSelector
Implements nodeSelector mapping from kubernetes pod's to firmament's taskdescriptor.
The network requirement for the netcost model will be defined in the podSpec as follows:

nodeSelector:
 networkRequirement: 1048
 disk: ssd

The value of networkRequirement under the nodeSelector will be updated to the 'ResourceRequest' of
the task descriptor and the rest of the value in the nodeSelector will be updated as LabelSelector in firmament.

Affected Modules: k8sclient

Change-Id: I7783435e3848f9305fecb3f50e88735857f0194c